Refactor hg_util into Galaxy.
